feat: Enhance Compomint core features and comprehensive documentation
This commit significantly enhances the Compomint core engine's capabilities
and updates the README with comprehensive documentation covering usage,
API, advanced features, internal workings, and best practices.

Key Changes:
- Core Functionality: Refined template syntax, data binding, event handling, i18n, lifecycle management.
- Installation & Basic Usage: Detailed guidance on various template definition methods and component creation/usage.
- Advanced Features: Template remapping, custom settings, built-in components (co-Ele, co-Element).
- Internal Workings: Explained template compilation, lazy evaluation, and scope architecture.
- Best Practices: Guidelines for performance optimization, code organization, and error handling.
- Example Applications: Added practical examples like Todo App and Theme Switcher.
- Test Suite: Documented the web-based test suite structure and usage (test-web/README.md).

This update empowers developers to better understand and leverage Compomint
for building robust web applications.
