+++ examples/greeting.esm.html
@@ -1,0 +1,82 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Compomint Core Greeting (ESM) Demo</title>
+  <!-- 
+      This example uses the ES Module version of Compomint.
+      Ensure the path to compomint.esm.js is correct.
+      It's expected to be in ../dist/ relative to this examples/ directory.
+    -->
+  <style>
+    body {
+      font-family: sans-serif;
+      padding: 20px;
+    }
+
+    .greeting {
+      border: 1px solid #ccc;
+      padding: 10px;
+      margin-top: 10px;
+      background-color: #f9f9f9;
+    }
+
+    .user-info {
+      color: blue;
+    }
+
+    .item {
+      list-style-type: square;
+    }
+  </style>
+</head>
+
+<body>
+  <h1>Compomint Core Greeting (ESM) Demo</h1>
+
+  <div id="app-container">
+    <!-- Components will be rendered here -->
+  </div>
+
+  <!-- Define Compomint Templates -->
+  <template id="my-greeting-template">
+    <div class="greeting">
+      <h2>##=data.title##</h2>
+      <p>Hello, <strong class="user-info">##=data.user.name##</strong>! You have ##=data.user.messages## messages.</p>
+      ##if(data.items && data.items.length > 0){##
+      <p>Your items:</p>
+      <ul>
+        ##for(let i = 0; i < data.items.length; i++){## <li class="item">##=data.items[i]##</li>
+          ##}##
+      </ul>
+      ##}##
+      <button data-co-event="##:{click: function() { alert('Button clicked for ' + data.user.name); }}##">Click
+        Me</button>
+    </div>
+  </template>
+
+  <script type="module">
+    import { compomint } from '../dist/compomint.esm.min.js';
+
+    document.addEventListener('DOMContentLoaded', function () {
+      compomint.addTmpls(document.documentElement);
+      const greetingRenderer = compomint.tmpl('my-greeting-template');
+      const greetingData = {
+        title: "Welcome to Compomint (ESM)!",
+        user: { name: "Alex (ESM)", messages: 5 },
+        items: ["Apples", "Bananas", "Cherries"]
+      };
+      const myGreetingComponent = greetingRenderer(greetingData);
+      const appContainer = document.getElementById('app-container');
+      if (appContainer && myGreetingComponent && myGreetingComponent.element) {
+        appContainer.appendChild(myGreetingComponent.element);
+      } else {
+        console.error("Could not render or append ESM component.");
+      }
+    });
+  </script>
+</body>
+
+</html>
