+++ src/compomint-core.ts
@@ -3,156 +3,14 @@
  * Code released under the MIT license
  */
 
-// Define interfaces for better type safety
-interface CompomintConfigs {
-  printExecTime?: boolean;
-  debug?: boolean;
-  throwError?: boolean;
-}
-
-interface TemplateMeta {
-  renderingFunc: RenderingFunction;
-  source?: string;
-  templateText?: string;
-  elements?: Set<any>; // Assuming elements is a Set, adjust type as needed
-}
-
-interface LazyScope {
-  [key: string]: any[]; // General structure for lazy scope arrays
-  elementPropsArray: any[];
-  namedElementArray: string[];
-  elementRefArray: ((target: Element) => void)[];
-  elementLoadArray: { loadFunc: Function; customData: any }[];
-  eventArray: { eventFunc: Function | Record<string, Function>; $parent: any; customData: any }[][];
-  elementArray: { childTarget: any; nonblocking: boolean | number }[];
-  lazyEvaluateArray: ((data: Record<string, any>) => void)[];
-}
-
-interface TemplateRule {
-  pattern: RegExp;
-  exec: (...args: any[]) => string;
-  lazyExec?: (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element) => void;
-  attacher?: (
-    self: any,
-    data: Record<string, any>,
-    lazyScope: LazyScope,
-    component: ComponentScope,
-    wrapper: DocumentFragment | Element,
-    elementTrigger: Element,
-    eventFunc: Function | Record<string, Function>,
-    eventData: Record<string, any>
-  ) => void;
-  trigger?: (target: Element, eventName: string) => void;
-}
-
-
-interface TemplateConfig {
-  rules: Record<string, TemplateRule>;
-  keys: {
-    dataKeyName: string,
-    statusKeyName: string,
-    componentKeyName: string,
-    i18nKeyName: string,
-  };
-}
-
-interface ComponentScope {
-  tmplId: string;
-  _id: string;
-  element: HTMLElement | TemplateElement;
-  wrapperElement?: Element;
-  parentComponent?: ComponentScope;
-  status: Record<string, any>;
-  data: Record<string, any>;
-  trigger?: Record<string, Record<string, () => void>>;
-  [key: string]: any; // Allow arbitrary properties for named elements etc.
-
-  replace(newComponent: ComponentScope | Element): void;
-  remove(spacer?: boolean): Element | TemplateElement | Comment;
-  appendTo(parentElement: Element): ComponentScope;
-  release(): void;
-  render(newData: Record<string, any>): ComponentScope;
-  refresh(reflashData: Record<string, any>): ComponentScope;
-  reflash(reflashData: Record<string, any>): ComponentScope; // Alias for refresh
-
-  // Optional lifecycle hooks
-  beforeRemove?(): void;
-  afterRemove?(): void;
-  beforeAppendTo?(): void;
-  afterAppendTo?(): void;
-  beforeRefresh?(): void;
-  afterRefresh?(): void;
-}
-
-type RenderingFunction = (
-  data?: any,
-  wrapperElement?: Element | ((component: ComponentScope) => void),
-  callback?: (component: ComponentScope) => void,
-  baseComponent?: Partial<ComponentScope>
-) => ComponentScope;
-
-interface Tools {
-  escapeHtml: {
-    escape: (str: string) => string;
-    unescape: (str: string) => string;
-  };
-  genId: (tmplId: string) => string;
-  genElement: (tagName: string, attrs?: Record<string, any> | string | Node | (string | Node)[], ...children: (string | Node)[]) => Element;
-  props: (...propsObjects: Record<string, any>[]) => string;
-  liveReloadSupport?: (component: ComponentScope) => void;
-  addTmpl: (tmplId: string, element: Element | string, tmplSettings?: Partial<TemplateConfig>) => RenderingFunction;
-  addTmpls: (source: Element | string, removeInnerTemplate?: boolean | Partial<TemplateConfig>, tmplSettings?: Partial<TemplateConfig>) => Element | TemplateElement;
-  addTmplByUrl: (importData: string | any[] | { url: string; option?: Record<string, any> }, option?: Record<string, any> | (() => void), callback?: () => void) => void;
-}
-
-interface CompomintGlobal {
-  configs: CompomintConfigs;
-  tmplCache: Map<string, TemplateMeta>;
-  templateConfig: TemplateConfig;
-  tools: Tools;
-  i18n: Record<string, any>;
-  template: (tmplId: string, templateText: string, tmplSettings?: Partial<TemplateConfig>) => RenderingFunction;
-  remapTmpl: (json: Record<string, string>) => void;
-  tmpl: (tmplId: string) => RenderingFunction | null;
-  addTmpl: (tmplId: string, element: Element | string, tmplSettings?: Partial<TemplateConfig>) => RenderingFunction;
-  addTmpls: (source: Element | string, removeInnerTemplate?: boolean | Partial<TemplateConfig>, tmplSettings?: Partial<TemplateConfig>) => Element | TemplateElement;
-  addTmplByUrl: (importData: string | any[] | { url: string; option?: Record<string, any> }, option?: Record<string, any> | (() => void), callback?: () => void) => void;
-  addI18n: (fullKey: string, i18nObj: Record<string, any>) => void;
-  addI18ns: (i18nObjs: Record<string, any>) => void;
-}
-
-
-/*
-declare global {
-  namespace globalThis {
-    var compomint: compomintGlobal;
-    var tmpl: Record<string, any>; // Global namespace for template functions
-  }
-  interface Node {
-    prototype: Node;
-    new(): Node;
-  }
-  interface CharacterData {
-    prototype: CharacterData;
-    new(): CharacterData;
-  }
-  interface DocumentType {
-    prototype: DocumentType;
-    new(): DocumentType;
-  }
-  interface TemplateElement extends HTMLTemplateElement { // Define TemplateElement if not globally available
-    prototype: HTMLTemplateElement;
-    new(): HTMLTemplateElement;
-  }
-}
-  */
-
-
-// Ensure TemplateElement is defined for environments like older JSDOM
-interface TemplateElement extends HTMLTemplateElement { }
-
-
-// Extend Window interface
+import { CompomintGlobal, ComponentScope, LazyScope, RenderingFunction, TemplateConfig, TemplateMeta, TemplateRule, Tools, TemplateElement } from "./type";
+import { defaultTemplateConfig } from "./default-template-config";
+import { applyBuiltInTemplates } from "./built-in-templates"
+import {
+  firstElementChild,
+  childElementCount,
+  cleanNode,
+} from "./utils";
 
 
 // Polyfill for Object.assign
@@ -209,20 +67,15 @@ if (typeof Object.assign != "function") {
   });
 })("isConnected" in window.Node.prototype);
 
-
 const compomint = ({} as CompomintGlobal);
 const tmpl = {} as Record<string, any>;
 const tools = (compomint.tools = compomint.tools || {} as Tools);
 const configs = (compomint.configs = Object.assign({ printExecTime: false, debug: false, throwError: true }, compomint.configs));
-const domParser = new DOMParser()
-
 const cachedTmpl = (compomint.tmplCache = compomint.tmplCache || new Map<string, TemplateMeta>());
 if (!cachedTmpl.has("anonymous")) {
   cachedTmpl.set("anonymous", { elements: new Set() } as TemplateMeta); // Cast to TemplateMeta
 }
 const isSupportTemplateTag = "content" in document.createElement("template");
-
-
 
 const noMatch = /(.)^/;
 const escapes: Record<string, string> = {
@@ -240,522 +93,8 @@ const escapes: Record<string, string> = {
 
 const escaper = /\>( |\n)+\<|\>( |\n)+|( |\n)+\<|\\|'|\r|\n|\t|\u2028|\u2029/g;
 
-const firstElementChild = function (ele: Element | DocumentFragment): Element | null {
-  if (ele.firstElementChild) return ele.firstElementChild;
-  const children = ele.childNodes;
-  for (let i = 0, size = children.length; i < size; i++) {
-    if (children[i] instanceof Element) {
-      return children[i] as Element;
-    }
-  }
-  return null;
-};
-
-const childNodeCount = function (ele: Element | DocumentFragment): number {
-  return (
-    ele.childElementCount ||
-    Array.prototype.filter.call(ele.childNodes, function (child: Node) {
-      return child instanceof Node;
-    }).length
-  );
-};
-
-const childElementCount = function (ele: Element | DocumentFragment): number {
-  return (
-    ele.childElementCount ||
-    Array.prototype.filter.call(ele.childNodes, function (child: Node) {
-      return child instanceof Element;
-    }).length
-  );
-};
-
-const cleanNode = function (node: Node): void {
-  for (let n = 0; n < node.childNodes.length; n++) {
-    const child = node.childNodes[n];
-    if (
-      child.nodeType === 8 || // Comment node
-      (child.nodeType === 3 && !/\S/.test(child.nodeValue || '')) // Text node with only whitespace
-    ) {
-      node.removeChild(child);
-      n--; // Adjust index after removal
-    } else if (child.nodeType === 1) {
-      // Element node
-      cleanNode(child); // Recurse
-    }
-  }
-};
-const stringToElement = function (str: string | number): Node {
-  if (typeof str === 'number' || !isNaN(Number(str))) {
-    return document.createTextNode(String(str));
-  } else if (typeof str === 'string') {
-    try {
-      const doc = domParser.parseFromString(str, "text/html");
-      const body = doc.body;
-      if (body.childNodes.length === 1) {
-        return body.firstChild!;
-      } else {
-        const fragment = document.createDocumentFragment();
-        while (body.firstChild) {
-          fragment.appendChild(body.firstChild);
-        }
-        return fragment;
-      }
-    } catch (e) {
-      return document.createTextNode(str);
-    }
-  } else {
-    return document.createTextNode('');
-  }
-};
-
-compomint.templateConfig = {
-  rules: {
-    style: {
-      pattern: /(\<style id=[\s\S]+?\>[\s\S]+?\<\/style\>)/g,
-      exec: function (style: string): string {
-        const dumy = document.createElement("template");
-        dumy.innerHTML = style;
-        const styleNode = (dumy.content || dumy).querySelector("style");
-        if (!styleNode || !styleNode.id) return ""; // Skip if no style node or ID
-        const oldStyleNode = document.getElementById(styleNode.id);
-        if (oldStyleNode) oldStyleNode.parentNode?.removeChild(oldStyleNode);
-        document.head.appendChild(styleNode);
-        return "";
-      },
-    },
-    commentArea: {
-      pattern: /##\*([\s\S]+?)##/g,
-      exec: function (commentArea: string): string {
-        return ``;
-      },
-    },
-    preEvaluate: {
-      pattern: /##!([\s\S]+?)##/g,
-      exec: function (preEvaluate: string, tmplId: string): string {
-        try {
-          new Function("tmplId", preEvaluate)(tmplId);
-        } catch (e: any) {
-          if (configs.throwError) {
-            console.error(`Template preEvaluate error in "${tmplId}", ${e.name}: ${e.message}`);
-            throw e;
-          } else {
-            console.warn(`Template preEvaluate error in "${tmplId}", ${e.name}: ${e.message}`);
-          }
-        }
-        return ``;
-      },
-    },
-    interpolate: {
-      pattern: /##=([\s\S]+?)##/g,
-      exec: function (interpolate: string): string {
-        const interpolateSyntax = `typeof (interpolate)=='function' ? (interpolate)() : (interpolate)`;
-        return (`';\n(() => {let __t, interpolate=${interpolate};\n__p+=((__t=(${interpolateSyntax}))==null ? '' : String(__t) );})();\n__p+='`); // Ensure string conversion
-      },
-    },
-    escape: {
-      pattern: /##-([\s\S]+?)##/g,
-      exec: function (escape: string): string {
-        const escapeSyntax = `compomint.tools.escapeHtml.escape(typeof (escape)=='function' ? (escape)() : (escape))`;
-        return (`';\n(() => {let __t, escape=${escape};\n__p+=((__t=(${escapeSyntax}))==null ? '' : String(__t) );})();\n__p+='`); // Ensure string conversion before escape
-      },
-    },
-    elementProps: {
-      pattern: /data-co-props="##:([\s\S]+?)##"/g,
-      exec: function (props: string): string {
-        const source = `';\nvar eventId = (__lazyScope.elementPropsArray.length);\n__p+='data-co-props="'+eventId+'"';\n
-__lazyScope.elementPropsArray[eventId] = ${props};\n__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        lazyScope.elementPropsArray.forEach(function (props: Record<string, any> | null, eventId: number) {
-          if (!props) return;
-          const $elementTrigger = wrapper.querySelector<Element>(
-            `[data-co-props="${eventId}"]`
-          );
-          if (!$elementTrigger) return;
-          delete ($elementTrigger as HTMLElement).dataset.coProps;
-          Object.keys(props).forEach(function (key: string) {
-            $elementTrigger.setAttribute(key, String(props[key])); // Ensure value is string
-          });
-        });
-      }
-    },
-    namedElement: {
-      pattern: /data-co-named-element="##:([\s\S]+?)##"/g,
-      exec: function (key: string): string {
-        const source = `';\nvar eventId = (__lazyScope.namedElementArray.length);\n__p+='data-co-named-element="'+eventId+'"';\n
-__lazyScope.namedElementArray[eventId] = ${key};\n__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        lazyScope.namedElementArray.forEach(function (key: string, eventId: number) {
-          const $elementTrigger = wrapper.querySelector<Element>(
-            `[data-co-named-element="${eventId}"]`
-          );
-          if (!$elementTrigger) {
-            if (configs.debug) console.warn(`Named element target not found for ID ${eventId} in template ${component.tmplId}`);
-            return;
-          }
-          delete ($elementTrigger as HTMLElement).dataset.coNamedElement;
-          component[key] = $elementTrigger;
-        });
-      },
-    },
-    elementRef: {
-      pattern: /data-co-element-ref="##:([\s\S]+?)##"/g,
-      exec: function (key: string): string {
-        const source = `';\nvar eventId = (__lazyScope.elementRefArray.length);\n__p+='data-co-element-ref="'+eventId+'"';
-var ${key} = null;\n__lazyScope.elementRefArray[eventId] = function(target) {${key} = target;};\n__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        lazyScope.elementRefArray.forEach(function (func: (target: Element) => void, eventId: number) {
-          const $elementTrigger = wrapper.querySelector<Element>(
-            `[data-co-element-ref="${eventId}"]`
-          );
-          if (!$elementTrigger) {
-            if (configs.debug) console.warn(`Element ref target not found for ID ${eventId} in template ${component.tmplId}`);
-            return;
-          }
-          delete ($elementTrigger as HTMLElement).dataset.coElementRef;
-          func.call($elementTrigger, $elementTrigger);
-        });
-      },
-    },
-    elementLoad: {
-      pattern: /data-co-load="##:([\s\S]+?)##"/g,
-      exec: function (elementLoad: string): string {
-        const elementLoadSplitArray = elementLoad.split("::");
-        const source = `';\nlet eventId = (__lazyScope.elementLoadArray.length);\n__p+='data-co-load="'+eventId+'"';
-__lazyScope.elementLoadArray[eventId] = {loadFunc: ${elementLoadSplitArray[0]}, customData: ${elementLoadSplitArray[1]}};\n__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        lazyScope.elementLoadArray.forEach(function (elementLoad: { loadFunc: Function; customData: Record<string, any> }, eventId: number) {
-          const $elementTrigger = wrapper.querySelector<Element>(
-            `[data-co-load="${eventId}"]`
-          );
-          if (!$elementTrigger) {
-            if (configs.debug) console.warn(`Element load target not found for ID ${eventId} in template ${component.tmplId}`);
-            return;
-          }
-          delete ($elementTrigger as HTMLElement).dataset.coLoad;
-          try {
-            elementLoad.loadFunc.call(
-              $elementTrigger,
-              $elementTrigger,
-              {
-                "data": data,
-                "element": $elementTrigger,
-                "customData": elementLoad.customData,
-                "component": component,
-              },
-            );
-          } catch (e: any) {
-            console.error(`Error executing elementLoad function for ID ${eventId} in template ${component.tmplId}:`, e, elementLoad.loadFunc);
-            if (configs.throwError) throw e;
-          }
-        });
-      },
-    },
-    event: {
-      pattern: /data-co-event="##:([\s\S]+?)##"/g,
-      exec: function (event: string): string {
-        const eventStrArray = event.split(":::");
-        let source = `';\n(() => {let eventId = (__lazyScope.eventArray.length);\n__p+='data-co-event="'+eventId+'"';\n`;
-        const eventArray: string[] = [];
-        for (let i = 0, size = eventStrArray.length; i < size; i++) {
-          const eventSplitArray = eventStrArray[i].split("::");
-          eventArray.push(
-            `{eventFunc: ${eventSplitArray[0]}, $parent: this, customData: ${eventSplitArray[1]}}`
-          );
-        }
-        source += `__lazyScope.eventArray[eventId] = [${eventArray.join(",")}];})()\n__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        const self = this as TemplateRule; // Cast self to TemplateSettings
-        const attacher = self.attacher;
-        if (!attacher) return; // Guard against missing attacher
-
-        lazyScope.eventArray.forEach(function (selectedArray: { eventFunc: Function | Record<string, Function>; $parent: any; customData: any }[], eventId: number) {
-          const $elementTrigger = wrapper.querySelector<Element>(
-            `[data-co-event="${eventId}"]`
-          );
-          if (!$elementTrigger) {
-            if (configs.debug) console.warn(`Event target not found for ID ${eventId} in template ${component.tmplId}`);
-            return;
-          }
-          delete ($elementTrigger as HTMLElement).dataset.coEvent;
-          for (let i = 0, size = selectedArray.length; i < size; i++) {
-            const selected = selectedArray[i];
-            if (selected.eventFunc) {
-              if (Array.isArray(selected.eventFunc)) {
-                selected.eventFunc.forEach(function (func: Function) {
-                  attacher(
-                    self, data, lazyScope, component, wrapper, $elementTrigger,
-                    func,
-                    selected
-                  );
-                });
-              } else {
-                attacher(
-                  self, data, lazyScope, component, wrapper, $elementTrigger,
-                  selected.eventFunc,
-                  selected
-                );
-              }
-            }
-          }
-        });
-      },
-      trigger: function (target: Element, eventName: string): void {
-        const customEvent = new Event(eventName, {
-          bubbles: true,
-          cancelable: true
-        });
-        target.dispatchEvent(customEvent);
-      },
-      attacher: function (
-        self: any, // Type properly if possible
-        data: Record<string, any>,
-        lazyScope: LazyScope,
-        component: ComponentScope,
-        wrapper: DocumentFragment | Element,
-        $elementTrigger: Element,
-        eventFunc: Function | Record<string, Function>,
-        eventData: Record<string, any>
-      ): void {
-        const trigger = self.trigger;
-        const $childTarget = firstElementChild(wrapper);
-        const $targetElement =
-          childElementCount(wrapper) == 1 ? $childTarget : null;
-
-        if (!eventFunc) {
-          return;
-        }
-
-        const eventFuncParams: [Element, Event | Element | null, { data: Record<string, any>; customData: Record<string, any>; element: Element; componentElement: Element | null | undefined; component: ComponentScope }] = [
-          $elementTrigger,
-          null,
-          {
-            "data": data,
-            "customData": eventData.customData,
-            "element": $elementTrigger,
-            "componentElement": $targetElement || $childTarget?.parentElement,
-            "component": component,
-          },
-        ];
-
-        if (typeof eventFunc === 'function') {
-          $elementTrigger.addEventListener("click", function (event: Event) {
-            event.stopPropagation();
-            eventFuncParams[1] = event;
-            try {
-              eventFunc.call(...eventFuncParams);
-            } catch (e: any) {
-              console.error(`Error in event handler for template ${component.tmplId}:`, e, eventFunc);
-              if (configs.throwError) throw e;
-            }
-          });
-          return;
-        }
-
-        const eventMap = eventFunc as Record<string, Function>;
-        const triggerName = eventMap.triggerName as unknown as string | undefined; // Optional key to store trigger functions
-        if (triggerName) {
-          component.trigger = component.trigger || {};
-          component.trigger[triggerName] = {};
-        }
-
-        Object.keys(eventMap).forEach(function (eventType: string) {
-          if (eventType == "load") {
-            eventFuncParams[1] = $elementTrigger;
-            try {
-              eventMap[eventType].call(...eventFuncParams);
-            } catch (e: any) {
-              console.error(`Error in 'load' event handler for template ${component.tmplId}:`, e, eventMap[eventType]);
-              if (configs.throwError) throw e;
-            }
-            return;
-          } else if (eventType == "namedElement") {
-            component[eventMap[eventType] as unknown as string] = $elementTrigger;
-            return;
-          } else if (eventType == "triggerName") {
-            return;
-          }
-
-          $elementTrigger.addEventListener(eventType, function (event: Event) {
-            event.stopPropagation();
-            eventFuncParams[1] = event;
-            try {
-              eventMap[eventType].call(...eventFuncParams);
-            } catch (e: any) {
-              console.error(`Error in '${eventType}' event handler for template ${component.tmplId}:`, e, eventMap[eventType]);
-              if (configs.throwError) throw e;
-            }
-          });
-
-          if (triggerName && trigger) {
-            component.trigger![triggerName][eventType] = function () {
-              trigger($elementTrigger, eventType);
-            };
-          }
-        });
-      },
-    },
-    element: {
-      pattern: /##%([\s\S]+?)##/g,
-      exec: function (target: string): string {
-        const elementSplitArray = target.split("::");
-        const source = `';\n(() => {
-let elementId = (__lazyScope.elementArray.length);
-__p+='<template data-co-tmpl-element-id="'+elementId+'"></template>';
-__lazyScope.elementArray[elementId] = {childTarget: ${elementSplitArray[0]}, nonblocking: ${(elementSplitArray[1] || false)}};})();
-__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        const self = this as TemplateRule; // Cast self
-
-        lazyScope.elementArray.forEach(function (ele: { childTarget: any; nonblocking: boolean | number }, elementId: number) {
-          const childTarget = ele.childTarget;
-          const nonblocking = ele.nonblocking;
-          const $tmplElement = wrapper.querySelector<TemplateElement>(
-            `template[data-co-tmpl-element-id="${elementId}"]`
-          );
-
-          if (!$tmplElement) {
-            if (configs.debug) console.warn(`Element insertion placeholder not found for ID ${elementId} in template ${component.tmplId}`);
-            return;
-          }
-          if (!$tmplElement.parentNode) {
-            if (configs.debug) console.warn(`Element insertion placeholder for ID ${elementId} in template ${component.tmplId} has no parent.`);
-            return;
-          }
-
-          const doFunc = function () {
-            if (!$tmplElement || !$tmplElement.parentNode) {
-              if (configs.debug) console.warn(`Placeholder for ID ${elementId} removed before insertion in template ${component.tmplId}.`);
-              return;
-            }
-
-            try {
-              if (childTarget instanceof Array) {
-                const docFragment = document.createDocumentFragment();
-                childTarget.forEach(function (child: any) {
-                  if (!child) return;
-                  const childElement = child.element || child;
-                  let nodeToAppend: Node | null = null;
-
-                  if (typeof childElement === "string" || typeof childElement === "number") {
-                    nodeToAppend = stringToElement(childElement);
-                  } else if (typeof childElement === "function") {
-                    nodeToAppend = stringToElement(childElement());
-                  } else if (childElement instanceof Node) {
-                    nodeToAppend = childElement;
-                  } else {
-                    if (configs.debug) console.warn(`Invalid item type in element array for ID ${elementId}, template ${component.tmplId}:`, childElement);
-                    return;
-                  }
-
-                  if (child.beforeAppendTo) { try { child.beforeAppendTo(); } catch (e) { console.error("Error in beforeAppendTo (array item):", e); } }
-                  if (nodeToAppend) docFragment.appendChild(nodeToAppend);
-                });
-
-                $tmplElement.parentNode.replaceChild(docFragment, $tmplElement);
-
-                childTarget.forEach(function (child: any) {
-                  if (child && child.afterAppendTo) {
-                    setTimeout(() => {
-                      try { child.afterAppendTo(); } catch (e) { console.error("Error in afterAppendTo (array item):", e); }
-                    }, 0);
-                  }
-                });
-
-              } else if (typeof childTarget === "string" || typeof childTarget === "number") {
-                $tmplElement.parentNode.replaceChild(
-                  stringToElement(childTarget),
-                  $tmplElement
-                );
-              } else if (typeof childTarget === "function") {
-                $tmplElement.parentNode.replaceChild(
-                  stringToElement(childTarget()),
-                  $tmplElement
-                );
-              } else if (childTarget && (childTarget.element || childTarget) instanceof Node) {
-                const childScope = childTarget as ComponentScope; // Assume it might be a scope
-                const childElement = childScope.element || childScope;
-
-                if (childScope.beforeAppendTo) { try { childScope.beforeAppendTo(); } catch (e) { console.error("Error in beforeAppendTo:", e); } }
-                $tmplElement.parentNode.replaceChild(childElement, $tmplElement);
-                if (childScope.afterAppendTo) {
-                  setTimeout(() => {
-                    try {
-                      if (childScope.afterAppendTo) childScope.afterAppendTo();
-                    } catch (e) { console.error("Error in afterAppendTo:", e); }
-                  }, 0);
-                }
-                if (childScope.tmplId) {
-                  childScope.parentComponent = component;
-                }
-
-              } else {
-                if (configs.debug) console.warn(`Invalid target for element insertion ID ${elementId}, template ${component.tmplId}:`, childTarget);
-                $tmplElement.parentNode.removeChild($tmplElement);
-              }
-            } catch (e: any) {
-              console.error(`Error during element insertion for ID ${elementId}, template ${component.tmplId}:`, e);
-              if (configs.throwError) throw e;
-              if ($tmplElement && $tmplElement.parentNode) {
-                try { $tmplElement.parentNode.removeChild($tmplElement); } catch (removeError) { /* Ignore */ }
-              }
-            }
-          }; // end doFunc
-
-          (nonblocking == undefined || nonblocking === false)
-            ? doFunc()
-            : setTimeout(doFunc, typeof nonblocking === 'number' ? nonblocking : 0);
-
-        }); // end forEach
-      }
-    },
-    lazyEvaluate: {
-      pattern: /###([\s\S]+?)##/g,
-      exec: function (lazyEvaluate: string): string {
-        const source = `';\n__lazyScope.lazyEvaluateArray.push(function(data) {${lazyEvaluate}});\n__p+='`;
-        return source;
-      },
-      lazyExec: function (data: Record<string, any>, lazyScope: LazyScope, component: ComponentScope, wrapper: DocumentFragment | Element): void {
-        const $childTarget = firstElementChild(wrapper);
-        const $targetElement =
-          childElementCount(wrapper) == 1 ? $childTarget : null;
-        lazyScope.lazyEvaluateArray.forEach(function (selectedFunc: (data: Record<string, any>) => void, idx: number) {
-          try {
-            selectedFunc.call($targetElement || wrapper, data); // Use wrapper if multiple elements
-          } catch (e: any) {
-            console.error(`Error in lazyEvaluate block ${idx} for template ${component.tmplId}:`, e, selectedFunc);
-            if (configs.throwError) throw e;
-          }
-        });
-        return;
-      },
-    },
-    evaluate: {
-      pattern: /##([\s\S]+?)##/g,
-      exec: (evaluate: string): string => {
-        return "';\n" + evaluate + "\n__p+='";
-      },
-    },
-  },
-  keys: {
-    dataKeyName: "data",
-    statusKeyName: "status",
-    componentKeyName: "component",
-    i18nKeyName: "i18n",
-  }
-
-} as TemplateConfig; // Cast to TemplateSettings
+// set default template config
+compomint.templateConfig = defaultTemplateConfig(configs, compomint);
 
 const escapeHtml = (function () {
   const escapeMap: Record<string, string> = {
@@ -1323,17 +662,13 @@ const safeTemplate = function (source: Element | string): Element | TemplateElem
   return template;
 };
 
-const addTmpl = (compomint.addTmpl = function (tmplId: string, element: Element | string, templateConfig?: Partial<TemplateConfig>): RenderingFunction {
+const addTmpl: CompomintGlobal["addTmpl"] = (compomint.addTmpl = function (tmplId, element, templateConfig) {
   let templateText = element instanceof Element ? element.innerHTML : String(element);
   templateText = escapeHtml.unescape(templateText.replace(/<!---|--->/gi, ""));
   return templateBuilder(tmplId, templateText, templateConfig);
 });
 
-const addTmpls = (compomint.addTmpls = function (
-  source: Element | string,
-  removeInnerTemplate?: boolean | Partial<TemplateConfig>,
-  templateConfig?: Partial<TemplateConfig>
-): Element | TemplateElement {
+const addTmpls: CompomintGlobal["addTmpls"] = (compomint.addTmpls = function (source, removeInnerTemplate, templateConfig) {
   if (typeof removeInnerTemplate !== "boolean" && templateConfig == undefined) {
     templateConfig = removeInnerTemplate as Partial<TemplateConfig> | undefined;
     removeInnerTemplate = false;
@@ -1364,11 +699,7 @@ const addTmpls = (compomint.addTmpls = function (
   return container;
 });
 
-const addTmplByUrl = (compomint.addTmplByUrl = function compomint_addTmplByUrl(
-  importData: string | any[] | { url: string; option?: Record<string, any> },
-  option?: Record<string, any> | (() => void),
-  callback?: Function | (() => void)
-): void {
+const addTmplByUrl: CompomintGlobal["addTmplByUrl"] = (compomint.addTmplByUrl = function compomint_addTmplByUrl(importData, option, callback) {
   if (!callback && typeof option === "function") {
     callback = option;
     option = {};
@@ -1668,24 +999,8 @@ tools.props = function (...propsObjects: Record<string, any>[]): string {
   return propStrArray.join(" ");
 };
 
-// Add built-in components
-addTmpl("co-Ele", `##%compomint.tools.genElement(data[0], data[1])##`);
-addTmpl(
-  "co-Element",
-  `##
-  data.tag = data.tag || 'div';
-  ##
-  &lt;##=data.tag##
-    ##=data.id ? 'id="' + (data.id === true ? component._id : data.id) + '"' : ''##
-    data-co-props="##:data.props##"
-    data-co-event="##:data.event##"&gt;
-    ##if (typeof data.content === "string") {##
-    ##=data.content##
-    ##} else {##
-      ##%data.content##
-    ##}##
-  &lt;/##=data.tag##&gt;`
-);
+// Add built-in template
+applyBuiltInTemplates(addTmpl);
 
 
 export { compomint, tmpl, CompomintGlobal };
