+++ test-web/index.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -12,42 +13,51 @@
             padding: 20px;
             line-height: 1.6;
         }
+
         h1 {
             text-align: center;
             color: #333;
             margin-bottom: 30px;
         }
+
         pre {
             background-color: #f5f5f5;
             padding: 15px;
             border-radius: 4px;
             overflow: auto;
         }
+
         .test-results {
             background-color: #f9f9f9;
             border-left: 4px solid #4CAF50;
             padding: 15px;
             margin-top: 25px;
         }
+
         .test-status {
             font-weight: bold;
         }
+
         .success {
             color: #4CAF50;
         }
+
         .failure {
             color: #F44336;
         }
+
         .test-group {
             background-color: #e9e9e9;
             padding: 10px;
             margin-bottom: 10px;
         }
+
         .test-item {
             padding: 5px;
             margin: 5px 0;
             border-bottom: 1px solid #ddd;
         }
+
         #run-tests {
             display: block;
             background-color: #4CAF50;
@@ -59,9 +69,11 @@
             margin: 20px auto;
             font-size: 1rem;
         }
+
         #run-tests:hover {
             background-color: #45a049;
         }
+
         #console-output {
             height: 300px;
             overflow-y: auto;
@@ -73,44 +85,44 @@
         }
     </style>
 </head>
+
 <body>
     <h1>Compomint Test Suite</h1>
-    
+
     <p>This page tests the template syntax and HTML element functionality of the Compomint library.</p>
-    
+
     <button id="run-tests">Run Tests</button>
-    
+
     <div class="test-results">
         <div id="test-summary">
             <h2>Test Results</h2>
             <p>No tests have been run yet.</p>
         </div>
-        
+
         <h3>Console Output:</h3>
         <div id="console-output"></div>
     </div>
-    
+
     <!-- Test template -->
     <template id="test-template">
         <div>
             <h2>##=data.title##</h2>
             <p>##=data.content##</p>
             ##if(data.showButton){##
-                <button data-co-event="##:{click: function() { alert('Button was clicked!'); }}##">##=data.buttonText##</button>
+            <button
+                data-co-event="##:{click: function() { alert('Button was clicked!'); }}##">##=data.buttonText##</button>
             ##}##
         </div>
     </template>
-    
-    <!-- Load Compomint library -->
-    <script src="../src/compomint-core.js"></script>
-    
+
     <!-- Load test scripts -->
     <script src="compomint-test-suite.js"></script>
     <script src="template-tests.js"></script>
     <script src="html-element-tests.js"></script>
     <script src="run-tests.js"></script>
-    
-    <script>
+
+    <script type="module">
+        import { compomint } from '../dist/compomint.esm.min.js';
 
         // Display console output on screen
         const consoleOutput = document.getElementById('console-output');
@@ -119,10 +131,10 @@
             error: console.error,
             warn: console.warn
         };
-        
-        console.log = function(...args) {
+
+        console.log = function (...args) {
             originalConsole.log.apply(console, args);
-            const message = args.map(arg => 
+            const message = args.map(arg =>
                 typeof arg === 'object' ? JSON.stringify(arg) : String(arg)
             ).join(' ');
             const logElement = document.createElement('div');
@@ -130,10 +142,10 @@
             consoleOutput.appendChild(logElement);
             consoleOutput.scrollTop = consoleOutput.scrollHeight;
         };
-        
-        console.error = function(...args) {
+
+        console.error = function (...args) {
             originalConsole.error.apply(console, args);
-            const message = args.map(arg => 
+            const message = args.map(arg =>
                 typeof arg === 'object' ? JSON.stringify(arg) : String(arg)
             ).join(' ');
             const logElement = document.createElement('div');
@@ -142,15 +154,15 @@
             consoleOutput.appendChild(logElement);
             consoleOutput.scrollTop = consoleOutput.scrollHeight;
         };
-        
+
         // Set up test run button
-        document.getElementById('run-tests').addEventListener('click', function() {
+        document.getElementById('run-tests').addEventListener('click', function () {
             consoleOutput.innerHTML = '';
             CompomintTest.assertions = 0;
             CompomintTest.passed = 0;
             CompomintTest.failed = 0;
-            
-            runAllTests();
+
+            runAllTests(compomint);
 
             // Update test results summary
             const testSummary = document.getElementById('test-summary');
@@ -162,13 +174,14 @@
                     Failed: <span class="failure">${CompomintTest.failed}</span>
                 </p>
                 <p>
-                    ${CompomintTest.failed === 0 
-                        ? '<span class="success">✓ All tests passed!</span>' 
-                        : `<span class="failure">✗ ${CompomintTest.failed} tests failed.</span>`
-                    }
+                    ${CompomintTest.failed === 0
+                    ? '<span class="success">✓ All tests passed!</span>'
+                    : `<span class="failure">✗ ${CompomintTest.failed} tests failed.</span>`
+                }
                 </p>
             `;
         });
     </script>
 </body>
+
 </html>
