+++ test-web/run-tests.js
@@ -1,15 +1,15 @@
 // Run all tests
-function runAllTests() {
+function runAllTests(compomint) {
   console.log('===== Compomint Tests Started =====');
-  
-  runTemplateTests();
-  runHtmlElementTests();
-  
+
+  runTemplateTests(compomint);
+  runHtmlElementTests(compomint);
+
   console.log('\n===== Test Summary =====');
   console.log(`Total Assertions: ${CompomintTest.assertions}`);
   console.log(`Passed: ${CompomintTest.passed}`);
   console.log(`Failed: ${CompomintTest.failed}`);
-  
+
   if (CompomintTest.failed === 0) {
     console.log('\n✓ All tests passed!');
   } else {
@@ -18,7 +18,7 @@ function runAllTests() {
 }
 
 // Run tests on page load
-document.addEventListener('DOMContentLoaded', runAllTests);
+document.addEventListener('DOMContentLoaded', () => { runAllTests(compomint) });
 
 // Export modules (for Node.js environment)
 if (typeof module !== 'undefined' && module.exports) {
