+++ test/compomint-core-built-in-templates.test.ts
@@ -1,16 +1,16 @@
 /**
  * @jest-environment jsdom
  */
-import { compomint, tmpl } from './compomint-core';
+
+import { describe, it, expect, beforeAll, beforeEach, afterEach, jest } from '@jest/globals';
+import { compomint, tmpl, CompomintGlobal } from "../src/compomint-core";
+
 
 // --- Built-in Component Tests ---
 describe('Built-in Components', () => {
-  let tools;
-  let configs;
+
 
   beforeAll(() => {
-    tools = compomint.tools;
-    configs = compomint.configs;
     if (!compomint) {
       throw new Error("Compomint library not loaded correctly into JSDOM environment.");
     }
@@ -25,8 +25,9 @@ describe('Built-in Components', () => {
     compomint.i18n = {};
 
     // Reset debug/error configs if needed (or set specific values for tests)
-    configs.debug = false;
-    configs.throwError = true; // Default to throwing errors for easier debugging in tests
+    compomint.i18n = {};
+    compomint.configs.debug = false;
+    compomint.configs.throwError = true; // Default to throwing errors for easier debugging in tests
 
     // Reset unique ID counter if possible (requires modifying source or using a test build)
     // For now, we assume IDs increment across tests, which is usually fine.
@@ -38,7 +39,7 @@ describe('Built-in Components', () => {
     it('co-Ele should render an element using genElement', () => {
       // Note: genElement expects 'className' not 'class' for the class attribute
 
-      const component = compomint.tmpl('co-Ele')(['button', { id: 'btn1', class: 'btn', 'data-tmpl-name': 'element-test', innerHTML: "<span>click</span" }]);
+      const component = compomint.tmpl('co-Ele')!(['button', { id: 'btn1', class: 'btn', 'data-tmpl-name': 'element-test', innerHTML: "<span>click</span" }]);
       expect(component.element.tagName).toBe('BUTTON');
       expect(component.element.outerHTML).toBe('<button id="btn1" class="btn" data-tmpl-name="element-test"><span>click</span></button>');
     });
@@ -48,17 +49,17 @@ describe('Built-in Components', () => {
   describe('co-Element', () => {
 
     it('co-Element should render a default div', () => {
-      const component = compomint.tmpl('co-Element')({});
+      const component = compomint.tmpl('co-Element')!({});
       expect(component.element.tagName).toBe('DIV');
     });
 
     it('co-Element should render the specified tag', () => {
-      const component = compomint.tmpl('co-Element')({ tag: 'span' });
+      const component = compomint.tmpl('co-Element')!({ tag: 'span' });
       expect(component.element.tagName).toBe('SPAN');
     });
 
     it('co-Element should render with id, class, and style', () => {
-      const component = compomint.tmpl('co-Element')({ id: 'myId', props: { class: 'my-class', style: 'color: red;', 'data-tmpl-name': 'element-test' } });
+      const component = compomint.tmpl('co-Element')!({ id: 'myId', props: { class: 'my-class', style: 'color: red;', 'data-tmpl-name': 'element-test' } });
       expect(component.element.id).toBe('myId');
       expect(component.element.className).toBe('my-class');
       expect(component.element.style.color).toBe('red');
@@ -66,21 +67,21 @@ describe('Built-in Components', () => {
     });
 
     it('co-Element should render with component._id when data.id is true', () => {
-      const component = compomint.tmpl('co-Element')({ id: true });
+      const component = compomint.tmpl('co-Element')!({ id: true });
       // The template uses `scope._id`, which likely refers to `component._id` in the execution context
       expect(component.element.id).toBe(component._id);
       expect(component._id).toMatch(/co-Element\d+/); // Check if the ID format is correct
     });
 
     it('co-Element should render with string content', () => {
-      const component = compomint.tmpl('co-Element')({ content: 'Hello World' });
+      const component = compomint.tmpl('co-Element')!({ content: 'Hello World' });
       expect(component.element.innerHTML).toContain('Hello World');
     });
 
     it('co-Element should render with element/component content', () => {
       const innerSpan = document.createElement('span');
       innerSpan.textContent = 'Inner';
-      const component = compomint.tmpl('co-Element')({ content: innerSpan });
+      const component = compomint.tmpl('co-Element')!({ content: innerSpan });
       expect(component.element.innerHTML).toBe('<span>Inner</span>');
     });
